// frontend/src/app/model/DateComptable.ts
export class DateComptable {
    datecompt_codtable: number;
	datecompt_typtable: number;
	datecompt_typcentral: number;
	datecompt_codejournal: Date;
	datecompt_libellejourn: Date;
	datecompt_datejourn: Date;
	datecompt_datemens: string;
	datecompt_dateexercice: string;
	datecompt_cloture: Date;
	datecompt_clotproch: string;
	datecompt_exerciceouv: string;
	datecompt_datemodif: Date;
	datecompt_coduser: string;
}